python-services/rag-query-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector Database
chromadb==0.4.18

# LangChain
langchain==0.0.340
langchain-community==0.0.1

# Service Discovery
nacos-sdk-python==0.1.9

# Cache
redis==5.0.1

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-zipkin==1.21.0

# Utilities
httpx==0.25.2
python-multipart==0.0.6
